Refactored to CustomComparatorsClass
